(feat) app: basic resource gathering (#8)
Unit moves to a resource and harvests it. The unit doesn't return the harvested resource yet (since there isn't a town hall or any other resource building yet). Attack scripts were modified to be an action subclass, and now actions are handled via an action queue on units. Harvest is also an action subclass.